Platform/Sgi: refactor StandaloneMM platform description file
The RD-N2 platform has a different memory map from that of the other
platforms supported under the SgiPkg. To enable the use of StandaloneMM
as a secure partition on RD-N2 platform, refactor the existing
StandaloneMM platform description file. The differing portions are split
into two different files and the rest of the  platform description file
is converted into a include file.

Signed-off-by: Sayanta Pattanayak <sayanta.pattanayak@arm.com>